+++ src/reviews.js
@@ -5,19 +5,24 @@
 'use strict';
 
 (function() {
-  var getReviewElement = require('./reviews_template');
-  var filter = require('./reviews_filter');
-  var getPageReview = require('./reviews_paigination');
-
   var filterList = document.querySelector('.reviews-filter');
   var filterItem = filterList['reviews'];
   var reviewContainer = document.querySelector('.reviews-list');
+  var template = document.querySelector('#review-template');
   var sectionReviews = document.querySelector('.reviews');
   var buttonMore = document.querySelector('.reviews-controls-more');
   var defaultFilter = 'reviews-all';
+  var cloneElement;
   var reviews;
 
   filterList.classList.add('invisible');
+
+// 1 Находим нужный элемент в разметке. Для поддержки старыми браузерами (усли браузер не поддерживает тег template) добавляем else
+  if ('content' in template) {
+    cloneElement = template.content.querySelector('.review');
+  } else {
+    cloneElement = template.querySelector('.review');
+  }
 
   /** @constant {number} */
   var IMAGE_LOAD_TIMEOUT = 10000;
@@ -28,7 +33,60 @@
   /** @type {Array.<Object>} */
   var filteredReviews = [];
 
+  /** @constant {number} */
+  var PAGE_SIZE = 3;
+
+  /** @constant {number} */
   var pageNumber = 0;
+
+// 2
+  /**
+   * @param {Object} data
+   * @param {HTMLElement} container
+   * @return {HTMLElement}
+   */
+  function getReviewElement(data, container) {
+    // Клонируем шаблонный элемент и заполняем элементы отеля данными из объекта data
+    var element = cloneElement.cloneNode(true);
+    element.querySelector('.review-text').textContent = data.description;
+
+    // вставлеяем элемент в DOM
+    container.appendChild(element);
+
+    var backgroundImage = new Image();
+    var backgroundLoadTimeout;
+
+    backgroundImage.addEventListener('load', function() {
+      clearTimeout(backgroundLoadTimeout);
+      element.replaceChild(backgroundImage, element.children[0]);
+      backgroundImage.classList.add('review-author');
+      backgroundImage.alt = 'Отзыв пользоватедя ' + data.author.name + ' об игре Code & Magick';
+      backgroundImage.title = data.author.name;
+      backgroundImage.width = 124;
+      backgroundImage.height = 124;
+    });
+
+    backgroundImage.addEventListener('error', function() {
+      element.classList.add('review-load-failure');
+    });
+
+    backgroundImage.src = data.author.picture;
+
+    backgroundLoadTimeout = setTimeout(function() {
+      backgroundImage.src = '';
+      element.classList.add('review-load-failure');
+    }, IMAGE_LOAD_TIMEOUT);
+
+
+    var reviewRatingBlock = element.querySelector('.review-rating');
+    var rating = data.rating;
+
+    var transRating = ['one', 'two', 'three', 'four', 'five'];
+
+    reviewRatingBlock.classList.add('review-rating-' + transRating[rating - 1]);
+
+    return element;
+  }
 
   function getReviewList(callback) {
     var xhr = new XMLHttpRequest();
@@ -54,10 +112,16 @@
     xhr.send();
   }
 
+  function isNextPageAvailable(reviewItem, pageNum, pageSize) {
+    return pageNum < Math.floor(reviewItem.length / pageSize);
+  }
+
   function btnMoreActive() {
     buttonMore.addEventListener('click', function() {
-      pageNumber++;
-      renderReview(filteredReviews, pageNumber);
+      if (isNextPageAvailable(reviews, pageNumber, PAGE_SIZE)) {
+        pageNumber++;
+        renderReview(filteredReviews, pageNumber);
+      }
     });
   }
 
@@ -72,11 +136,24 @@
    * @param {boolean=} replace
    * */
   function renderReview(data, page, replace) {
-    var pageReview = getPageReview(data, page, replace);
+    if (replace) {
+      reviewContainer.innerHTML = '';
+    }
+
+    var from = page * PAGE_SIZE;
+    var to = from + PAGE_SIZE;
+
+    var pageReview = data.slice(from, to);
 
     pageReview.forEach(function(reviewItem) {
       getReviewElement(reviewItem, reviewContainer);
     });
+
+    if (isNextPageAvailable(reviews, pageNumber, PAGE_SIZE)) {
+      buttonMore.classList.remove('invisible');
+    } else {
+      buttonMore.classList.add('invisible');
+    }
   }
 
   filterList.addEventListener('change', function() {
@@ -85,7 +162,46 @@
 
   function setActiveFilter(id) {
     pageNumber = 0;
-    filteredReviews = filter(id, reviews);
+    filteredReviews = reviews.slice(0);
+    switch (id) {
+      case 'reviews-all':
+        break;
+      // показывает список отзывов, оставленных за две недели, отсортированных по убыванию даты
+      case 'reviews-recent':
+        filteredReviews = filteredReviews.sort(function(a, b) {
+          return b.date - a.date;
+        });
+        filteredReviews = filteredReviews.filter(function(a) {
+          var lastDate = new Date();
+          lastDate.setDate(lastDate.getDate() - 14);
+          return a.date > lastDate;
+        });
+        break;
+      //  с рейтингом не ниже 3, отсортированные по убыванию рейтинга
+      case 'reviews-good':
+        filteredReviews = filteredReviews.filter(function(a) {
+          return a.rating > 2;
+        });
+        filteredReviews = filteredReviews.sort(function(a, b) {
+          return b.rating - a.rating;
+        });
+        break;
+      //  с рейтингом не выше 2, отсортированные по возрастанию рейтинга
+      case 'reviews-bad':
+        filteredReviews = filteredReviews.filter(function(a) {
+          return a.rating < 3;
+        });
+        filteredReviews = filteredReviews.sort(function(a, b) {
+          return a.rating - b.rating;
+        });
+        break;
+      case 'reviews-popular':
+        //  отсортированные по убыванию оценки отзыва
+        filteredReviews = filteredReviews.sort(function(a, b) {
+          return b.review_usefulness - a.review_usefulness;
+        });
+        break;
+    }
 
     renderReview(filteredReviews, pageNumber, true);
   }
